backend/project/requirements.txt
fastapi==0.63.0
pydantic==1.7.3
starlette==0.13.6
PyYAML==5.4.1
click==7.1.2
h11==0.12.0
httptools==0.1.1
python-dotenv==0.15.0
uvicorn==0.13.3
uvloop==0.15.0 
watchgod==0.6 
websockets==8.1
Mako==1.1.4
MarkupSafe==1.1.1
SQLAlchemy==1.3.23
alembic==1.5.4
pytest==6.2.2
passlib==1.7.4
python-dotenv==0.15.0
python-jose==3.2.0
python-multipart==0.0.5